Conflicts:
	source/blockUpdaters/blockUpdater.cpp
	source/files.cpp
	source/world.h

+++ source/blockUpdaters/CactusUpdater.cpp
@@ -16,10 +16,10 @@ CactusUpdater::CactusUpdater()
 
 bool CactusUpdater::update(WorldObject &world, int x, int y, bool bg)
 {
-	int &blockBelowXY = bg ? world.bgblocks[x][y + 1] : world.blocks[x][y + 1];
+	short &blockBelowXY = bg ? world.bgblocks[x][y + 1] : world.blocks[x][y + 1];
 	if (blockBelowXY != SAND && blockBelowXY != CACTUS)
 	{
-		int &blockXY = bg ? world.bgblocks[x][y] : world.blocks[x][y];
+		short &blockXY = bg ? world.bgblocks[x][y] : world.blocks[x][y];
 		blockXY = AIR;
 		createItemMob(x, y, CACTUS);
 		return true;
@@ -29,11 +29,11 @@ bool CactusUpdater::update(WorldObject &world, int x, int y, bool bg)
 
 void CactusUpdater::chanceUpdate(WorldObject &world, int x, int y, bool bg)
 {
-	int &blockAboveXY = bg ? world.bgblocks[x][y - 1] : world.blocks[x][y - 1];
-	int &blockBelowXY = bg ? world.bgblocks[x][y + 1] : world.blocks[x][y + 1];
+	short &blockAboveXY = bg ? world.bgblocks[x][y - 1] : world.blocks[x][y - 1];
+	short &blockBelowXY = bg ? world.bgblocks[x][y + 1] : world.blocks[x][y + 1];
 	if (blockAboveXY == AIR && blockBelowXY == SAND)
 	{
-		int &blockXY = bg ? world.bgblocks[x][y] : world.blocks[x][y];
+		short &blockXY = bg ? world.bgblocks[x][y] : world.blocks[x][y];
 		blockAboveXY = blockXY;
 	}
 }
