+++ source/blockUpdaters/OakSaplingUpdater.cpp
@@ -14,7 +14,7 @@ void saplingChanceUpdate(WorldObject &world, int x, int y, bool bg)
 {
 	if (world.brightness[x][y + 1] < 2)
 	{
-		int &blockXY = bg ? world.bgblocks[x][y] : world.blocks[x][y];
+		short &blockXY = bg ? world.bgblocks[x][y] : world.blocks[x][y];
 		switch (blockXY)
 		{
 		case SAPLING_OAK:
@@ -33,8 +33,8 @@ void saplingChanceUpdate(WorldObject &world, int x, int y, bool bg)
 
 void saplingUpdate(WorldObject &world, int x, int y, bool bg)
 {
-	int &blockXY = bg ? world.bgblocks[x][y] : world.blocks[x][y];
-	int blockBelowXY = bg ? world.bgblocks[x][y + 1] : world.blocks[x][y + 1];
+	short &blockXY = bg ? world.bgblocks[x][y] : world.blocks[x][y];
+	short blockBelowXY = bg ? world.bgblocks[x][y + 1] : world.blocks[x][y + 1];
 
 	int requiredBlock = AIR;
 	switch (blockXY)
